# docs/openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "TalkShore API",
    "version": "1.0.0",
    "description": "REST API for TalkShore Voyage and the TalkShore Control Tower. Protected endpoints require a Firebase ID token in the Authorization bearer header.",
    "contact": {
      "name": "TalkShore API Team"
    }
  },
  "tags": [
    {
      "name": "Health",
      "description": "Deployment and dependency readiness."
    },
    {
      "name": "Authentication",
      "description": "Verified Firebase identity."
    },
    {
      "name": "Profiles",
      "description": "The authenticated user's TalkShore profile."
    }
  ],
  "servers": [
    {
      "url": "/api",
      "description": "Current deployment"
    }
  ],
  "components": {
    "securitySchemes": {
      "firebaseBearer": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "Firebase-ID-Token",
        "description": "Firebase ID token returned after successful Firebase Authentication login."
      }
    },
    "schemas": {
      "Profile": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "email": {
            "type": "string",
            "nullable": true,
            "format": "email"
          },
          "full_name": {
            "type": "string",
            "nullable": true
          },
          "role": {
            "type": "string",
            "enum": [
              "learner",
              "guide",
              "moderator",
              "admin"
            ]
          },
          "avatar_url": {
            "type": "string",
            "nullable": true,
            "format": "uri"
          },
          "native_language": {
            "type": "string",
            "nullable": true
          },
          "learning_language": {
            "type": "string",
            "nullable": true
          },
          "level": {
            "type": "string",
            "nullable": true
          },
          "goal": {
            "type": "string",
            "nullable": true
          },
          "days_per_week": {
            "type": "integer",
            "nullable": true,
            "minimum": 1,
            "maximum": 7
          },
          "banned": {
            "type": "boolean"
          },
          "verified_guide": {
            "type": "boolean"
          },
          "created_at": {
            "type": "string",
            "nullable": true,
            "format": "date-time"
          },
          "updated_at": {
            "type": "string",
            "nullable": true,
            "format": "date-time"
          }
        },
        "required": [
          "id",
          "email",
          "full_name",
          "role",
          "banned",
          "verified_guide",
          "created_at",
          "updated_at"
        ]
      },
      "RegistrationResponse": {
        "type": "object",
        "properties": {
          "data": {
            "type": "object",
            "properties": {
              "profile": {
                "$ref": "#/components/schemas/Profile"
              },
              "custom_token": {
                "type": "string"
              },
              "token_type": {
                "type": "string",
                "enum": [
                  "firebase_custom_token"
                ]
              }
            },
            "required": [
              "profile",
              "custom_token",
              "token_type"
            ]
          }
        },
        "required": [
          "data"
        ]
      },
      "ErrorResponse": {
        "type": "object",
        "properties": {
          "error": {
            "type": "object",
            "properties": {
              "code": {
                "type": "string"
              },
              "message": {
                "type": "string"
              },
              "details": {
                "nullable": true
              },
              "request_id": {
                "type": "string"
              }
            },
            "required": [
              "code",
              "message",
              "request_id"
            ]
          }
        },
        "required": [
          "error"
        ]
      },
      "RegistrationInput": {
        "type": "object",
        "properties": {
          "full_name": {
            "type": "string",
            "minLength": 2,
            "maxLength": 100,
            "example": "Ada Mensah"
          },
          "email": {
            "type": "string",
            "maxLength": 254,
            "format": "email",
            "example": "ada@example.com"
          },
          "password": {
            "type": "string",
            "minLength": 8,
            "maxLength": 128,
            "example": "strong-password"
          },
          "role": {
            "type": "string",
            "enum": [
              "learner",
              "guide"
            ],
            "example": "learner"
          }
        },
        "required": [
          "full_name",
          "email",
          "password",
          "role"
        ],
        "additionalProperties": false
      },
      "ProfileResponse": {
        "type": "object",
        "properties": {
          "data": {
            "$ref": "#/components/schemas/Profile"
          }
        },
        "required": [
          "data"
        ]
      },
      "ProfileInput": {
        "type": "object",
        "properties": {
          "full_name": {
            "type": "string",
            "minLength": 2,
            "maxLength": 100
          }
        },
        "required": [
          "full_name"
        ],
        "additionalProperties": false
      },
      "ProfileUpdate": {
        "type": "object",
        "properties": {
          "full_name": {
            "type": "string",
            "minLength": 2,
            "maxLength": 100
          },
          "avatar_url": {
            "type": "string",
            "nullable": true,
            "format": "uri"
          },
          "native_language": {
            "type": "string",
            "nullable": true,
            "minLength": 2,
            "maxLength": 60
          },
          "learning_language": {
            "type": "string",
            "nullable": true,
            "minLength": 2,
            "maxLength": 60
          },
          "level": {
            "type": "string",
            "nullable": true,
            "minLength": 2,
            "maxLength": 40
          },
          "goal": {
            "type": "string",
            "nullable": true,
            "maxLength": 500
          },
          "days_per_week": {
            "type": "integer",
            "nullable": true,
            "minimum": 1,
            "maximum": 7
          }
        },
        "additionalProperties": false
      }
    },
    "parameters": {}
  },
  "paths": {
    "/api/v1/health": {
      "get": {
        "tags": [
          "Health"
        ],
        "summary": "Check API health",
        "operationId": "getHealth",
        "responses": {
          "200": {
            "description": "API health status",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "data": {
                      "type": "object",
                      "properties": {
                        "status": {
                          "type": "string",
                          "enum": [
                            "ok",
                            "degraded"
                          ]
                        },
                        "version": {
                          "type": "string"
                        },
                        "environment": {
                          "type": "string"
                        },
                        "firebase_admin": {
                          "type": "string",
                          "enum": [
                            "configured",
                            "not_configured"
                          ]
                        }
                      },
                      "required": [
                        "status",
                        "version",
                        "environment",
                        "firebase_admin"
                      ]
                    }
                  },
                  "required": [
                    "data"
                  ]
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/auth/register": {
      "post": {
        "tags": [
          "Authentication"
        ],
        "summary": "Register a learner or guide",
        "description": "Creates a Firebase Authentication user and matching TalkShore profile. Guides are created with verified_guide set to false.",
        "operationId": "registerUser",
        "security": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RegistrationInput"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "User registered",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RegistrationResponse"
                }
              }
            }
          },
          "409": {
            "description": "An account already exists for this email address",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "Registration details are invalid",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "503": {
            "description": "Firebase Admin is not configured",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/auth/me": {
      "get": {
        "tags": [
          "Authentication"
        ],
        "summary": "Get the current authenticated user",
        "operationId": "getCurrentUser",
        "security": [
          {
            "firebaseBearer": []
          }
        ],
        "responses": {
          "200": {
            "description": "Current user",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProfileResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Profile not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/profiles/me": {
      "get": {
        "tags": [
          "Profiles"
        ],
        "summary": "Get my profile",
        "operationId": "getMyProfile",
        "security": [
          {
            "firebaseBearer": []
          }
        ],
        "responses": {
          "200": {
            "description": "Profile",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProfileResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Profile not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Profiles"
        ],
        "summary": "Create or return my profile",
        "operationId": "createMyProfile",
        "security": [
          {
            "firebaseBearer": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ProfileInput"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Existing profile",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProfileResponse"
                }
              }
            }
          },
          "201": {
            "description": "Profile created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProfileResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      },
      "patch": {
        "tags": [
          "Profiles"
        ],
        "summary": "Update my profile",
        "operationId": "updateMyProfile",
        "security": [
          {
            "firebaseBearer": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ProfileUpdate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Updated profile",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProfileResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Profile not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    }
  }
}